+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - ArcGIS REST JS v2.x is now used under the hood
 - ember-auto-import is now used to load ArcGIS REST JS
 
+### Removed
+- `geocode-service` `findLocationAddress` method.
+
 ## [1.15.1]
 ### Changed
 - Allow updating `defaultUserCreditAssignment` & `platformSSO` on portal